+++ lib/src/rules/ntlife.rs
@@ -52,7 +52,7 @@ impl_rule! {
     /// * `0b_10` means dead,
     /// * `0b_01` means alive,
     /// * `0b_00` means unknown.
-    pub struct NbhdDesc(usize);
+    pub struct NbhdDesc(u32);
 
     /// Non-totalistic Life-like rules.
     ///
